+++ lib/hachoir/field/timestamp.py
@@ -1,6 +1,6 @@
 from hachoir.core.tools import (humanDatetime, humanDuration,
                                 timestampUNIX, timestampMac32, timestampUUID60,
-                                timestampWin64, durationWin64)
+                                timestampWin64, durationWin64, durationMillisWin64)
 from hachoir.field import Bits, FieldSet
 from datetime import datetime
 
@@ -91,3 +91,16 @@ class TimedeltaWin64(GenericTimestamp):
     def createValue(self):
         value = Bits.createValue(self)
         return durationWin64(value)
+
+
+class TimedeltaMillisWin64(GenericTimestamp):
+
+    def __init__(self, parent, name, description=None):
+        GenericTimestamp.__init__(self, parent, name, 64, description)
+
+    def createDisplay(self):
+        return humanDuration(self.value)
+
+    def createValue(self):
+        value = Bits.createValue(self)
+        return durationMillisWin64(value)